+++ apple-codesign/CHANGELOG.md
@@ -89,6 +89,8 @@ Released on ReleaseDate.
 * Fixed a bug where attempting to sign Mach-O binaries having a `__TEXT` segment
   whose start offset was >0 resulted in a `Mach-O segment corruption` error.
   We can now properly sign such files. (#91)
+* `verify` command now errors if not given the path of a Mach-O binary.
+* `verify` command now prints a warning that its known to be buggy.
 * aws crates 0.53 -> 0.57.
 * bitflags 1.3 -> 2.0.
 * cryptographic-message-syntax 0.19 -> 0.25.